refactor with easyeditor
Former-commit-id: 29ec43a097dd9eac48af1277fd049dd81da2d0f9

+++ easycomplex/src/libcomplex/LoadFromJson.cpp
@@ -1,7 +1,12 @@
-#include <common/dev_config.h>
-
 #include "LoadFromJson.h"
 #include "Symbol.h"
+
+#include <ee/dev_config.h>
+#include <ee/SymbolSearcher.h>
+#include <ee/SymbolMgr.h>
+#include <ee/Exception.h>
+#include <ee/SpriteFactory.h>
+#include <ee/Sprite.h>
 
 namespace ecomplex
 {
@@ -11,10 +16,10 @@ void LoadFromJson::Load(const Json::Value& value, const std::string& dir, Symbol
 	complex->name = value["name"].asString();
 	complex->tag = value["tag"].asString();
 
-	complex->m_clipbox.xmin = value["xmin"].asInt();
-	complex->m_clipbox.xmax = value["xmax"].asInt();
-	complex->m_clipbox.ymin = value["ymin"].asInt();
-	complex->m_clipbox.ymax = value["ymax"].asInt();
+	complex->m_clipbox.xmin = static_cast<float>(value["xmin"].asInt());
+	complex->m_clipbox.xmax = static_cast<float>(value["xmax"].asInt());
+	complex->m_clipbox.ymin = static_cast<float>(value["ymin"].asInt());
+	complex->m_clipbox.ymax = static_cast<float>(value["ymax"].asInt());
 
 	complex->m_use_render_cache = value["use_render_cache"].asBool();
